refactor: camera docs
